NAV portal 수정
