+++ src/components/Nav/Nav.jsx
@@ -1,11 +1,43 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { NAV_ITEM_LIST } from './NAV_ITEM_LIST';
+import LoginModal from '../Login/LoginModal';
+import SearchModal from '../SearchModal/SearchModal';
+import ViewItem from '../ViewItem/ViewItem';
+import ModalPortal from '../../ModalPortal';
+
 import './Nav.scss';
 
-function Nav({ showTargetModal }) {
+function Nav() {
+  const [modalState, setModalState] = useState({
+    login: false,
+    search: false,
+    viewItem: false,
+  });
   const [fullName, setFullName] = useState(null);
   const [token, setToken] = useState(null);
+
+  const showTargetModal = modalName => {
+    if (!['login', 'search', 'viewItem'].includes(modalName)) {
+      return;
+    }
+
+    setModalState(prev => ({
+      ...prev,
+      [modalName]: true,
+    }));
+  };
+
+  const closeTargetModal = modalName => {
+    if (!['login', 'search', 'viewItem'].includes(modalName)) {
+      return;
+    }
+
+    setModalState(prev => ({
+      ...prev,
+      [modalName]: false,
+    }));
+  };
+
   useEffect(() => {
     const token = localStorage.getItem('token');
     const fullName = localStorage.getItem('fullName');
@@ -26,21 +58,24 @@ function Nav({ showTargetModal }) {
   return (
     <>
       <div className="navTop">
-        <Link to="/" className="navTopLeft" />
+        <div className="navTopLeft">
+          <p className="navTopLeftItem">조던</p>
+          <p className="navTopLeftItem">컨버스</p>
+        </div>
         <div className="navTopRight">
           <Link to="/item-list" className="navTopRightItem">
             고객센터
           </Link>
           {token && fullName ? (
             <>
-              <p className="navTopRightItem">{fullName}</p>
+              <p className="navTopRightItem">{fullName}</p> |
               <button className="navTopRightItem" onClick={handleLogout}>
                 로그아웃
               </button>
             </>
           ) : (
             <>
-              <Link to="/sign-up" className="navTopRightItem">
+              <Link to="/sign-in" className="navTopRightItem">
                 멤버 가입
               </Link>
               <button
@@ -56,24 +91,28 @@ function Nav({ showTargetModal }) {
       <div className="navBottom">
         <div className="navBottomLeft">
           <Link to="/" className="navBottomLeftItem">
-            JUST-DO-IT
+            나이키
           </Link>
         </div>
         <div
           className="navBottomCenter"
           onMouseEnter={() => showTargetModal('viewItem')}
         >
-          {NAV_ITEM_LIST.map(({ navItemKey, name }) => {
-            return (
-              <Link
-                key={navItemKey}
-                to="/item-list"
-                className="navBottomCenterItem"
-              >
-                {name}
-              </Link>
-            );
-          })}
+          <Link to="/item-list" className="navBottomCenterItem">
+            New Releases
+          </Link>
+          <Link to="/item-list" className="navBottomCenterItem">
+            Men
+          </Link>
+          <Link to="/item-list" className="navBottomCenterItem">
+            Women
+          </Link>
+          <Link to="/item-list" className="navBottomCenterItem">
+            Kids
+          </Link>
+          <Link to="/item-list" className="navBottomCenterItem">
+            Sale
+          </Link>
         </div>
         <div className="navBottomRight">
           <div className="navBottomRightSearch">
@@ -97,6 +136,21 @@ function Nav({ showTargetModal }) {
           </div>
         </div>
       </div>
+      {modalState.login && (
+        <ModalPortal>
+          <LoginModal closeTargetModal={closeTargetModal} />
+        </ModalPortal>
+      )}
+      {modalState.search && (
+        <ModalPortal>
+          <SearchModal closeTargetModal={closeTargetModal} />
+        </ModalPortal>
+      )}
+      {modalState.viewItem && (
+        <ModalPortal>
+          <ViewItem closeTargetModal={closeTargetModal} />
+        </ModalPortal>
+      )}
     </>
   );
 }
