+++ src/components/Nav/Nav.scss
@@ -9,7 +9,7 @@
     width: 150px;
     height: 80px;
     margin-left: 70px;
-    background-image: url(../../../public/image/navimage.png);
+    // background-image: url(../../../public/image/navimage.png);
     background-size: cover;
     background-position: center;
   }
